api/requirements.txt
flask 
flask_cors 
psycopg2-binary
gunicorn 
boto3
python-dotenv
